LPS-75321 Moves code from BaseLayoutDisplayContext to LayoutsAdminDisplayContext
